data-mining/requirements.txt
hydra-core==1.3
lightning
pandas
wandb
rich
torchmetrics